backend/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9
pydantic>=2.6.1
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
requests>=2.31.0
email-validator>=2.1.0
argon2-cffi>=21.5.0
google-auth>=2.38.0
redis>=5.0.0
gunicorn>=21.2.0
